refactor(reposicao): extrai AdminReposicaoAumentoDetail god-component (1465→738)
2º refactor da série de god-components (§10). Extrai os subcomponentes e
types/helpers de AdminReposicaoAumentoDetail.tsx para src/components/reposicao/aumentoDetail/.
Extração pura, behavior-preserving (zero mudança de lógica/JSX/queries):

- types.ts          — Aumento, Item, Mapeamento, SkuAfetado
- shared.ts         — EMPRESA, ESTADOS_LABEL, estadoBadgeClass, confiancaClass, diasEntre
- CancelarButton.tsx, ItemRow.tsx
- MapeamentoDialog.tsx — MapeamentoDialog + type FamiliaSelecionada (co-locado)
- SkuPickerDialog.tsx

A página vira o orquestrador (738 linhas: o componente principal de detalhe do aumento).
Import block enxuto (só o que o main usa). Revisado via codex review (extração fiel,
sem bug). typecheck:strict + tsc + 392/392 testes + lint 0.
