+++ src/pages/AdminReposicaoAumentoDetail.tsx
@@ -8,16 +8,12 @@ import {
   ArrowLeft,
   Loader2,
   Save,
-  Trash2,
   ExternalLink,
   Plus,
   CheckCircle2,
-  XCircle,
-  Layers,
   TrendingUp,
   Sparkles,
   Mail,
-  Search as SearchIcon,
 } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
@@ -25,14 +21,6 @@ import { Badge } from "@/components/ui/badge";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
-import { Checkbox } from "@/components/ui/checkbox";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 import {
   Tabs,
   TabsContent,
@@ -48,117 +36,15 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import {
-  Dialog,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogDescription,
-} from "@/components/ui/dialog";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@/components/ui/alert-dialog";
-import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
 } from "@/components/ui/tooltip";
-
-const EMPRESA = "OBEN";
-
-type Aumento = {
-  id: number;
-  empresa: string;
-  nome: string;
-  fornecedor_nome: string;
-  data_vigencia: string;
-  data_anuncio: string | null;
-  estado: string;
-  observacoes: string | null;
-  origem_arquivo_url: string | null;
-  origem_arquivo_tipo: string | null;
-  origem_email_assunto: string | null;
-  origem_email_remetente: string | null;
-  origem_email_data: string | null;
-  extracao_confianca: number | null;
-  extracao_observacoes: string | null;
-};
-
-type Item = {
-  id: number;
-  aumento_id: number;
-  categoria_fornecedor: string;
-  aumento_perc: number;
-  data_vigencia_especifica: string | null;
-  confirmado: boolean;
-  ativo: boolean;
-  observacoes: string | null;
-};
-
-type Mapeamento = {
-  id: number;
-  aumento_item_id: number;
-  familia_omie: string;
-  sku_codigo_omie_especifico: number | null;
-};
-
-type SkuAfetado = {
-  sku_codigo_omie: number;
-  sku_descricao: string | null;
-  familia: string | null;
-  categoria_fornecedor: string;
-  data_vigencia_efetiva: string;
-  aumento_perc: number;
-};
-
-const ESTADOS_LABEL: Record<string, string> = {
-  rascunho: "Rascunho",
-  ativo: "Ativo",
-  vigente: "Vigente",
-  expirado: "Expirado",
-  cancelado: "Cancelado",
-};
-
-function estadoBadgeClass(estado: string): string {
-  switch (estado) {
-    case "rascunho":
-      return "bg-status-warning/15 text-status-warning border-status-warning/30";
-    case "ativo":
-      return "bg-status-info/15 text-status-info border-status-info/30";
-    case "vigente":
-      return "bg-status-success/15 text-status-success border-status-success/30";
-    case "expirado":
-      return "bg-muted text-muted-foreground border-border";
-    case "cancelado":
-      return "bg-destructive/15 text-destructive border-destructive/30";
-    default:
-      return "";
-  }
-}
-
-function confiancaClass(c: number | null): string {
-  if (c === null) return "";
-  if (c < 0.5) return "bg-destructive/15 text-destructive border-destructive/30";
-  if (c <= 0.8)
-    return "bg-status-warning/15 text-status-warning border-status-warning/30";
-  return "bg-status-success/15 text-status-success border-status-success/30";
-}
-
-function diasEntre(data: string): number {
-  const target = new Date(data + "T00:00:00");
-  const today = new Date();
-  today.setHours(0, 0, 0, 0);
-  return Math.round((target.getTime() - today.getTime()) / (1000 * 60 * 60 * 24));
-}
+import { Aumento, Item, Mapeamento, SkuAfetado } from "@/components/reposicao/aumentoDetail/types";
+import { EMPRESA, ESTADOS_LABEL, estadoBadgeClass, confiancaClass, diasEntre } from "@/components/reposicao/aumentoDetail/shared";
+import { CancelarButton } from "@/components/reposicao/aumentoDetail/CancelarButton";
+import { ItemRow } from "@/components/reposicao/aumentoDetail/ItemRow";
 
 export default function AdminReposicaoAumentoDetail() {
   const { id: idParam } = useParams<{ id: string }>();
@@ -848,618 +734,5 @@ export default function AdminReposicaoAumentoDetail() {
         </div>
       </div>
     </TooltipProvider>
-  );
-}
-
-// ============= SUB-COMPONENTS =============
-
-function CancelarButton({ onConfirm }: { onConfirm: () => void }) {
-  return (
-    <AlertDialog>
-      <AlertDialogTrigger asChild>
-        <Button variant="outline" className="w-full text-destructive">
-          <XCircle className="h-4 w-4" /> Cancelar
-        </Button>
-      </AlertDialogTrigger>
-      <AlertDialogContent>
-        <AlertDialogHeader>
-          <AlertDialogTitle>Cancelar anúncio de aumento?</AlertDialogTitle>
-          <AlertDialogDescription>
-            Esta ação não pode ser desfeita. O anúncio ficará marcado como
-            cancelado.
-          </AlertDialogDescription>
-        </AlertDialogHeader>
-        <AlertDialogFooter>
-          <AlertDialogCancel>Voltar</AlertDialogCancel>
-          <AlertDialogAction onClick={onConfirm}>
-            Confirmar cancelamento
-          </AlertDialogAction>
-        </AlertDialogFooter>
-      </AlertDialogContent>
-    </AlertDialog>
-  );
-}
-
-function ItemRow({
-  item,
-  numFamilias,
-  onUpdate,
-  onDelete,
-  onMapeamentoChanged,
-}: {
-  item: Item;
-  numFamilias: number;
-  onUpdate: (patch: Partial<Item>) => void;
-  onDelete: () => void;
-  onMapeamentoChanged: () => void;
-}) {
-  const [categoria, setCategoria] = useState(item.categoria_fornecedor);
-  const [perc, setPerc] = useState(String(item.aumento_perc));
-  const [vig, setVig] = useState(item.data_vigencia_especifica ?? "");
-  const [mapDialogOpen, setMapDialogOpen] = useState(false);
-
-  useEffect(() => {
-    setCategoria(item.categoria_fornecedor);
-    setPerc(String(item.aumento_perc));
-    setVig(item.data_vigencia_especifica ?? "");
-  }, [item]);
-
-  return (
-    <>
-      <TableRow>
-        <TableCell>
-          <Input
-            value={categoria}
-            onChange={(e) => setCategoria(e.target.value)}
-            onBlur={() => {
-              if (categoria !== item.categoria_fornecedor) {
-                onUpdate({ categoria_fornecedor: categoria });
-              }
-            }}
-          />
-        </TableCell>
-        <TableCell>
-          <Input
-            type="number"
-            step="0.01"
-            value={perc}
-            onChange={(e) => setPerc(e.target.value)}
-            onBlur={() => {
-              const n = Number(perc);
-              if (!isNaN(n) && n !== item.aumento_perc) {
-                onUpdate({ aumento_perc: n });
-              }
-            }}
-          />
-        </TableCell>
-        <TableCell>
-          <Input
-            type="date"
-            value={vig}
-            onChange={(e) => setVig(e.target.value)}
-            onBlur={() => {
-              const v = vig || null;
-              if (v !== item.data_vigencia_especifica) {
-                onUpdate({ data_vigencia_especifica: v });
-              }
-            }}
-          />
-        </TableCell>
-        <TableCell>
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() => setMapDialogOpen(true)}
-            className="w-full justify-start"
-          >
-            <Layers className="h-3.5 w-3.5" />
-            <span className="text-xs">
-              {numFamilias > 0
-                ? `${numFamilias} ${numFamilias === 1 ? "família" : "famílias"}`
-                : "Mapear"}
-            </span>
-          </Button>
-        </TableCell>
-        <TableCell className="text-center">
-          <Checkbox
-            checked={item.confirmado}
-            onCheckedChange={(c) => onUpdate({ confirmado: c === true })}
-          />
-        </TableCell>
-        <TableCell>
-          <AlertDialog>
-            <AlertDialogTrigger asChild>
-              <Button variant="ghost" size="sm">
-                <Trash2 className="h-4 w-4 text-destructive" />
-              </Button>
-            </AlertDialogTrigger>
-            <AlertDialogContent>
-              <AlertDialogHeader>
-                <AlertDialogTitle>Remover categoria?</AlertDialogTitle>
-                <AlertDialogDescription>
-                  A categoria "{item.categoria_fornecedor}" será desativada.
-                </AlertDialogDescription>
-              </AlertDialogHeader>
-              <AlertDialogFooter>
-                <AlertDialogCancel>Voltar</AlertDialogCancel>
-                <AlertDialogAction onClick={onDelete}>
-                  Remover
-                </AlertDialogAction>
-              </AlertDialogFooter>
-            </AlertDialogContent>
-          </AlertDialog>
-        </TableCell>
-      </TableRow>
-
-      <MapeamentoDialog
-        open={mapDialogOpen}
-        onOpenChange={setMapDialogOpen}
-        item={item}
-        onSaved={() => {
-          setMapDialogOpen(false);
-          onMapeamentoChanged();
-        }}
-      />
-    </>
-  );
-}
-
-// ============= MAPEAMENTO DIALOG =============
-
-type FamiliaSelecionada = {
-  familia: string;
-  apenasEspecificos: boolean;
-  skusEspecificos: number[];
-};
-
-function MapeamentoDialog({
-  open,
-  onOpenChange,
-  item,
-  onSaved,
-}: {
-  open: boolean;
-  onOpenChange: (o: boolean) => void;
-  item: Item;
-  onSaved: () => void;
-}) {
-  const [busca, setBusca] = useState("");
-  const [selecionadas, setSelecionadas] = useState<FamiliaSelecionada[]>([]);
-  const [skuPickerFor, setSkuPickerFor] = useState<string | null>(null);
-  const [salvando, setSalvando] = useState(false);
-
-  // Lista de famílias do Omie
-  const { data: familiasOmie = [] } = useQuery({
-    queryKey: ["omie-familias-aumento"],
-    enabled: open,
-    queryFn: async () => {
-      const { data, error } = await supabase
-        .from("omie_products")
-        .select("familia")
-        .ilike("account", "oben")
-        .not("familia", "is", null)
-        .limit(5000);
-      if (error) throw error;
-      const set = new Set<string>();
-      ((data || []) as Array<{ familia: string | null }>).forEach((r) => r.familia && set.add(r.familia));
-      return Array.from(set).sort();
-    },
-  });
-
-  // Mapeamentos existentes para este item
-  const { data: existentes = [] } = useQuery({
-    queryKey: ["mapeamentos-item", item.id, open],
-    enabled: open,
-    queryFn: async () => {
-      const { data, error } = await supabase
-        .from("categoria_aumento_familia_mapeamento")
-        .select("familia_omie, sku_codigo_omie_especifico")
-        .eq("aumento_item_id", item.id);
-      if (error) throw error;
-      return (data || []) as unknown as Array<{
-        familia_omie: string;
-        sku_codigo_omie_especifico: number | null;
-      }>;
-    },
-  });
-
-  // Hidrata estado quando abre
-  useEffect(() => {
-    if (!open) return;
-    const grouped: Record<string, FamiliaSelecionada> = {};
-    for (const e of existentes) {
-      if (!grouped[e.familia_omie]) {
-        grouped[e.familia_omie] = {
-          familia: e.familia_omie,
-          apenasEspecificos: false,
-          skusEspecificos: [],
-        };
-      }
-      if (e.sku_codigo_omie_especifico !== null) {
-        grouped[e.familia_omie].apenasEspecificos = true;
-        grouped[e.familia_omie].skusEspecificos.push(
-          e.sku_codigo_omie_especifico,
-        );
-      }
-    }
-    setSelecionadas(Object.values(grouped));
-  }, [open, existentes]);
-
-  const familiasFiltradas = useMemo(() => {
-    const q = busca.trim().toLowerCase();
-    if (!q) return familiasOmie.slice(0, 30);
-    return familiasOmie.filter((f) => f.toLowerCase().includes(q)).slice(0, 50);
-  }, [busca, familiasOmie]);
-
-  const adicionarFamilia = (familia: string) => {
-    if (selecionadas.find((s) => s.familia === familia)) return;
-    setSelecionadas([
-      ...selecionadas,
-      { familia, apenasEspecificos: false, skusEspecificos: [] },
-    ]);
-    setBusca("");
-  };
-
-  const removerFamilia = (familia: string) => {
-    setSelecionadas(selecionadas.filter((s) => s.familia !== familia));
-  };
-
-  const toggleEspecificos = (familia: string, valor: boolean) => {
-    setSelecionadas(
-      selecionadas.map((s) =>
-        s.familia === familia
-          ? { ...s, apenasEspecificos: valor, skusEspecificos: valor ? s.skusEspecificos : [] }
-          : s,
-      ),
-    );
-    if (valor) setSkuPickerFor(familia);
-  };
-
-  const salvar = async () => {
-    setSalvando(true);
-    try {
-      // DELETE existentes
-      const { error: delErr } = await supabase
-        .from("categoria_aumento_familia_mapeamento")
-        .delete()
-        .eq("aumento_item_id", item.id);
-      if (delErr) throw delErr;
-
-      // INSERT novos
-      const inserts: Array<{
-        aumento_item_id: number;
-        familia_omie: string;
-        sku_codigo_omie_especifico: number | null;
-      }> = [];
-      for (const s of selecionadas) {
-        if (s.apenasEspecificos && s.skusEspecificos.length > 0) {
-          for (const sku of s.skusEspecificos) {
-            inserts.push({
-              aumento_item_id: item.id,
-              familia_omie: s.familia,
-              sku_codigo_omie_especifico: sku,
-            });
-          }
-        } else {
-          inserts.push({
-            aumento_item_id: item.id,
-            familia_omie: s.familia,
-            sku_codigo_omie_especifico: null,
-          });
-        }
-      }
-      if (inserts.length > 0) {
-        const { error: insErr } = await supabase
-          .from("categoria_aumento_familia_mapeamento")
-          .insert(inserts);
-        if (insErr) throw insErr;
-      }
-      toast.success("Mapeamento salvo");
-      onSaved();
-    } catch (e) {
-      toast.error(e instanceof Error ? e.message : "Erro ao salvar mapeamento");
-    } finally {
-      setSalvando(false);
-    }
-  };
-
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent className="max-w-3xl max-h-[85vh] overflow-y-auto">
-        <DialogHeader>
-          <DialogTitle>Mapear "{item.categoria_fornecedor}"</DialogTitle>
-          <DialogDescription>
-            Selecione as famílias do Omie afetadas por este aumento.
-            Opcionalmente restrinja a SKUs específicos por família.
-          </DialogDescription>
-        </DialogHeader>
-
-        <div className="space-y-4">
-          {/* Buscador */}
-          <div>
-            <Label>Adicionar família</Label>
-            <div className="relative mt-1.5">
-              <SearchIcon className="h-4 w-4 absolute left-3 top-1/2 -translate-y-1/2 text-muted-foreground" />
-              <Input
-                placeholder="Buscar família…"
-                className="pl-9"
-                value={busca}
-                onChange={(e) => setBusca(e.target.value)}
-              />
-            </div>
-            {busca.trim() && familiasFiltradas.length > 0 && (
-              <div className="mt-2 border rounded-md max-h-48 overflow-y-auto">
-                {familiasFiltradas.map((f) => (
-                  <button
-                    key={f}
-                    onClick={() => adicionarFamilia(f)}
-                    className="w-full text-left px-3 py-2 text-sm hover:bg-muted"
-                    disabled={!!selecionadas.find((s) => s.familia === f)}
-                  >
-                    {f}
-                    {selecionadas.find((s) => s.familia === f) && (
-                      <span className="text-xs text-muted-foreground ml-2">
-                        (já adicionada)
-                      </span>
-                    )}
-                  </button>
-                ))}
-              </div>
-            )}
-          </div>
-
-          {/* Lista de famílias selecionadas */}
-          <div className="space-y-2">
-            <Label>Famílias mapeadas ({selecionadas.length})</Label>
-            {selecionadas.length === 0 && (
-              <p className="text-sm text-muted-foreground italic">
-                Nenhuma família mapeada.
-              </p>
-            )}
-            {selecionadas.map((s) => {
-              const aplicaFamiliaInteira = !s.apenasEspecificos;
-              return (
-                <Card key={s.familia}>
-                  <CardContent className="pt-4 space-y-3">
-                    <div className="flex items-center justify-between gap-2">
-                      <span className="font-medium text-sm">{s.familia}</span>
-                      <Button
-                        variant="ghost"
-                        size="sm"
-                        onClick={() => removerFamilia(s.familia)}
-                      >
-                        <Trash2 className="h-4 w-4 text-destructive" />
-                      </Button>
-                    </div>
-
-                    {/* Toggle principal — default = família inteira */}
-                    <div
-                      className={`flex items-start gap-2 rounded-md border p-2.5 ${
-                        aplicaFamiliaInteira
-                          ? "border-primary/40 bg-primary/5"
-                          : "border-border"
-                      }`}
-                    >
-                      <Checkbox
-                        id={`fam-${s.familia}`}
-                        checked={aplicaFamiliaInteira}
-                        onCheckedChange={(c) =>
-                          toggleEspecificos(s.familia, !(c === true))
-                        }
-                        className="mt-0.5"
-                      />
-                      <div className="flex-1 min-w-0">
-                        <Label
-                          htmlFor={`fam-${s.familia}`}
-                          className="text-sm font-medium cursor-pointer"
-                        >
-                          Aplicar a TODA a família (todos os SKUs)
-                        </Label>
-                        <p className="text-xs text-muted-foreground mt-0.5">
-                          Salva uma única linha no banco. Desmarque para escolher SKUs específicos.
-                        </p>
-                      </div>
-                    </div>
-
-                    {!aplicaFamiliaInteira && (
-                      <div className="space-y-1.5">
-                        <p className="text-xs text-muted-foreground">
-                          {s.skusEspecificos.length} SKU
-                          {s.skusEspecificos.length === 1 ? "" : "s"} selecionado
-                          {s.skusEspecificos.length === 1 ? "" : "s"}
-                        </p>
-                        <Button
-                          variant="outline"
-                          size="sm"
-                          onClick={() => setSkuPickerFor(s.familia)}
-                        >
-                          Selecionar SKUs
-                        </Button>
-                      </div>
-                    )}
-                  </CardContent>
-                </Card>
-              );
-            })}
-          </div>
-        </div>
-
-        <DialogFooter>
-          <Button variant="ghost" onClick={() => onOpenChange(false)}>
-            Cancelar
-          </Button>
-          <Button onClick={salvar} disabled={salvando}>
-            {salvando && <Loader2 className="h-4 w-4 animate-spin" />}
-            Salvar mapeamento
-          </Button>
-        </DialogFooter>
-
-        {/* Sub-dialog: picker de SKUs */}
-        {skuPickerFor && (
-          <SkuPickerDialog
-            familia={skuPickerFor}
-            initialSelected={
-              selecionadas.find((s) => s.familia === skuPickerFor)?.skusEspecificos ?? []
-            }
-            onClose={() => setSkuPickerFor(null)}
-            onConfirm={(skus) => {
-              setSelecionadas(
-                selecionadas.map((s) =>
-                  s.familia === skuPickerFor
-                    ? { ...s, skusEspecificos: skus }
-                    : s,
-                ),
-              );
-              setSkuPickerFor(null);
-            }}
-          />
-        )}
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-function SkuPickerDialog({
-  familia,
-  initialSelected,
-  onClose,
-  onConfirm,
-}: {
-  familia: string;
-  initialSelected: number[];
-  onClose: () => void;
-  onConfirm: (skus: number[]) => void;
-}) {
-  const [selected, setSelected] = useState<Set<number>>(
-    new Set(initialSelected),
-  );
-  const [busca, setBusca] = useState("");
-
-  const { data: skus = [], isLoading } = useQuery({
-    queryKey: ["skus-familia", familia],
-    queryFn: async () => {
-      const { data, error } = await supabase
-        .from("omie_products")
-        .select("omie_codigo_produto, descricao")
-        .ilike("account", "oben")
-        .eq("familia", familia)
-        .order("descricao", { ascending: true })
-        .limit(2000);
-      if (error) throw error;
-      return (data || []) as Array<{
-        omie_codigo_produto: number;
-        descricao: string | null;
-      }>;
-    },
-  });
-
-  const filtrados = useMemo(() => {
-    const q = busca.trim().toLowerCase();
-    if (!q) return skus;
-    return skus.filter(
-      (s) =>
-        (s.descricao ?? "").toLowerCase().includes(q) ||
-        String(s.omie_codigo_produto).includes(q),
-    );
-  }, [skus, busca]);
-
-  const toggle = (codigo: number) => {
-    const next = new Set(selected);
-    if (next.has(codigo)) next.delete(codigo);
-    else next.add(codigo);
-    setSelected(next);
-  };
-
-  return (
-    <Dialog open onOpenChange={(o) => !o && onClose()}>
-      <DialogContent className="max-w-2xl max-h-[80vh]">
-        <DialogHeader>
-          <DialogTitle>SKUs de "{familia}"</DialogTitle>
-          <DialogDescription>
-            Selecione os SKUs específicos que recebem o aumento.
-          </DialogDescription>
-        </DialogHeader>
-        <div className="relative">
-          <SearchIcon className="h-4 w-4 absolute left-3 top-1/2 -translate-y-1/2 text-muted-foreground" />
-          <Input
-            placeholder="Buscar por código ou descrição…"
-            className="pl-9"
-            value={busca}
-            onChange={(e) => setBusca(e.target.value)}
-          />
-        </div>
-        <div className="flex items-center justify-between gap-2">
-          <p className="text-xs text-muted-foreground">
-            {filtrados.length} SKU{filtrados.length === 1 ? "" : "s"}
-            {busca.trim() ? " filtrado" : " na família"}
-            {filtrados.length !== 1 && busca.trim() ? "s" : ""}
-          </p>
-          <div className="flex gap-2">
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => {
-                const next = new Set(selected);
-                filtrados.forEach((s) => next.add(s.omie_codigo_produto));
-                setSelected(next);
-              }}
-              disabled={isLoading || filtrados.length === 0}
-            >
-              Selecionar todos
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => {
-                const next = new Set(selected);
-                filtrados.forEach((s) => next.delete(s.omie_codigo_produto));
-                setSelected(next);
-              }}
-              disabled={isLoading || filtrados.length === 0}
-            >
-              Desmarcar todos
-            </Button>
-          </div>
-        </div>
-        <div className="border rounded-md max-h-[400px] overflow-y-auto">
-          {isLoading ? (
-            <div className="p-4 text-sm text-muted-foreground flex items-center gap-2">
-              <Loader2 className="h-4 w-4 animate-spin" /> Carregando…
-            </div>
-          ) : filtrados.length === 0 ? (
-            <div className="p-4 text-sm text-muted-foreground">
-              Nenhum SKU encontrado.
-            </div>
-          ) : (
-            filtrados.map((sku) => (
-              <label
-                key={sku.omie_codigo_produto}
-                className="flex items-center gap-2 px-3 py-2 hover:bg-muted text-sm cursor-pointer border-b last:border-b-0"
-              >
-                <Checkbox
-                  checked={selected.has(sku.omie_codigo_produto)}
-                  onCheckedChange={() => toggle(sku.omie_codigo_produto)}
-                />
-                <span className="font-mono text-xs text-muted-foreground w-20 shrink-0">
-                  {sku.omie_codigo_produto}
-                </span>
-                <span className="truncate">{sku.descricao}</span>
-              </label>
-            ))
-          )}
-        </div>
-        <DialogFooter>
-          <p className="text-sm text-muted-foreground mr-auto">
-            {selected.size} selecionado{selected.size === 1 ? "" : "s"}
-          </p>
-          <Button variant="ghost" onClick={onClose}>
-            Cancelar
-          </Button>
-          <Button onClick={() => onConfirm(Array.from(selected))}>
-            Confirmar seleção
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
   );
 }
